replace text of links in source posts
This replacement only happens if the text of the link is the same as its
href. So links with customized texts are not replaced.

To do this I had to add a reverse many2many relation in discussion so
'references' is renamed to 'sources' and 'targets' is added. Both are
included by default in the requests, this way the title of the cross
referenced discussions are always available both ways.

An optional discussionId can be displayed next to the discussion link
but this will have to wait for the admin panel to be added.