Load discussion details from the API
Now that Links to target discussions are rendered from the backend, we
don't need to include the 'target' relation in the serializer.
Also, in order to prevent from leaking infos about hidden discussions,
it is best not to include any of the related discussions in the
discussion's serializer.

This automatically fixes the EventPost part of #10. The CommentPost will
be implemented in the backend in another commit.

This is a little bit more network intensive, as more requests have to be
made with discussions that have a lot of relations. But this is more
acceptable now that we cache in-flight requests to the API to
deduplicate them.