Fallback gracefully to unknown discussion rendering if request is null
As it seems a lot of extension do not set the parameter when calling
Formatter->render(), we fallback to unknown discussion rendering when
request is null.

We keep logging an error, as this issue should be solved in these
extensions.

See also <https://github.com/the-turk/flarum-diff/pull/35> for an
example of how to fix this issue.